refactored getprojcets usecase
