+++ src/test/kotlin/data/datasources/CsvParserTest.kt
@@ -5,7 +5,7 @@ import data.datasources.parser.CsvParser
 import data.datasources.parser.TaskCsvParser
 import logic.entities.Task
 import logic.exception.PlanMateException
-import logic.usecases.task.testFactory.CreateTaskTestFactory
+import logic.usecases.testFactory.CreateTaskTestFactory
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
